- Many new character tables for the database - Moved Organization tables into their own folder (simply for organizational purposes) - Removed User tables as they are now part of Characters - Lots of work on addressing code analysis issues brought up by Microsoft's code analysis - Fixed some bit extension methods to work with long values

+++ SmaugCS.Data/Templates/ObjectTemplate.cs
@@ -223,7 +223,7 @@ namespace SmaugCS.Data.Templates
                 Spells.Add(spell.ToLower());
         }
 
-        public void AddAffect(int type, int duration, int modifier, int location, int flags)
+        public void AddAffect(long type, int duration, int modifier, int location, int flags)
         {
             AffectData newAffect = new AffectData
             {
